will move common function to 'mock-helpers' module
- generateNames
- generateDate
- rename 'generateDescription' to 'generateText'
